+++ src/test/java/com/vaadin/ui/tabs/tests/TabTest.java
@@ -14,9 +14,11 @@
  * the License.
  */
 
-package com.vaadin.ui.tabs;
+package com.vaadin.ui.tabs.tests;
 
 import org.junit.Test;
+
+import com.vaadin.ui.tabs.Tab;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.*;
