+++ src/test/java/com/vaadin/ui/tabs/tests/TabsTest.java
@@ -14,11 +14,14 @@
  * the License.
  */
 
-package com.vaadin.ui.tabs;
+package com.vaadin.ui.tabs.tests;
 
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
+
+import com.vaadin.ui.tabs.Tab;
+import com.vaadin.ui.tabs.Tabs;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
